Move httpapi package into internal/codeintel/uploads/transport.
