+++ internal/codeintel/uploads/transport/httpapi/auth/github_test.go
