+++ internal/codeintel/uploads/transport/httpapi/auth/gitlab.go
