+++ internal/codeintel/uploads/transport/httpapi/auth/gitlab_test.go
