+++ internal/codeintel/uploads/transport/httpapi/auth/iface.go
