+++ internal/codeintel/uploads/transport/httpapi/auth/middleware.go
