+++ internal/codeintel/uploads/transport/httpapi/auth/mocks_test.go
