+++ internal/codeintel/uploads/transport/httpapi/auth/util.go
